requirements.txt
requests
python-dotenv
tabulate
httpx
rich
SQLAlchemy
psycopg2-binary
pandas

# StaffSpy (if pip install fails, install from its repo as per its README)
#pip install -U "staffspy[browser]"
